+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Telegram-bot (KtGram) Changelog
 
+## 9.3.1
+
+* Added `skipPersist` to `Transition.JumpTo` to give control over persisting.
+* Fixed parsing of parameters with default arg parser.
+
 ## 9.3.0
 
 * Covered Tg API [`9.5`](https://core.telegram.org/bots/api-changelog#march-1-2026).